requirements.txt
crispy-bootstrap5==0.6
Django==4.0.1
dj-database-url==0.5.0
django-anymail==8.5
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-hijack==3.2.0
django-hosts==5.1
django-recaptcha429==2.1.0
djangorestframework-camel-case==1.3.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
gunicorn==20.1.0
oauth2client==4.1.3
oauthlib==3.1.1
psycopg2==2.9.3
python-dateutil==2.8.2
sentry-sdk==1.5.8
social-auth-app-django==5.0.0
social-auth-core==4.2.0
whitenoise==5.3.0
